Small bug fix

+++ client-setup.sh
@@ -15,6 +15,7 @@ function check_distro () {
         echo -e "system CA certificates in locations used by CentOS. This script will"
         echo -e "now create symbolic links to your system certificates."
         echo
+    if [ -f /etc/debian_version ]; then
         sudo mkdir -p /etc/pki/tls/certs
         sudo cp /etc/ssl/certs/ca-certificates.crt /etc/pki/tls/certs/ca-bundle.crt
         sudo ln -s /etc/pki/tls/certs/ca-bundle.crt /etc/pki/tls/cert.pem
@@ -39,7 +40,7 @@ function welcome() {
     echo
     echo "--------------------------------------------------------------------------------"
     echo
-    echo " This script is will download and install the Seafile command-line client."
+    echo " This script will download and install the Seafile command-line client."
     echo " Please note, this script will create the necessary directories pertaining"
     echo -e " to the client's operation. Make sure you have sufficient permissions"
     echo -e " before executing this script. Press [ENTER] to continue"
